Added confirmation box before deletion

+++ src/app/advert/messaging-home/messaging-home.component.ts
@@ -8,6 +8,7 @@ import { RouterExtensions } from "nativescript-angular/router";
 import { ObservableArray } from "tns-core-modules/data/observable-array";
 //import for app settings
 import * as appSettings from "tns-core-modules/application-settings";
+import * as dialogs from "tns-core-modules/ui/dialogs";
 @Component({
     selector: 'ns-messaging-home',
     templateUrl: './messaging-home.component.html',
@@ -53,18 +54,21 @@ export class MessagingHomeComponent implements OnInit, OnDestroy {
                 if(deletechatresult){
                     this.deletechat = deletechatresult;
  
-                    if(this.deletechat.responseStatusCode === 200 && this.deletechat.chatdeleted === true){
+                    if(this.deletechat.responseStatusCode === 200 && this.deletechat.chatposted === true){
                         this.advertServ.initializeActiveChats();
+                        this.deletechatResultSub.unsubscribe();
                         TNSFancyAlert.showSuccess("Success!", "Chat Successfully Deleted!", "Close")
                         this.deleteChat = false;
                     } else if (this.deletechat.responseStatusCode === 500){
                         TNSFancyAlert.showError("Connection error", "An internal error has occured.", "Close");
                         this.deleteChat = false;
+                        this.deletechatResultSub.unsubscribe();
                     }
     
                     else {
                         TNSFancyAlert.showError("Error", "An Error has been recieved, please contact support." , "Close");
                         this.deleteChat = false;
+                        this.deletechatResultSub.unsubscribe();
                     }
                     
                 }
@@ -77,8 +81,24 @@ export class MessagingHomeComponent implements OnInit, OnDestroy {
     onItemSelected(args :ListViewEventData): void {
         if(this.deleteChat == true){
             const tappedActivechatItem = args.view.bindingContext;
-            this.advertServ.deleteChat(tappedActivechatItem.chatid);
-            //send to service
+            dialogs.confirm({
+                title: "Please confirm deletion",
+                message: "Are you sure that you wish to delete this chat?",
+                okButtonText: "Yes",
+                cancelButtonText: "No",
+            }).then(result => {
+                // result argument is boolean
+                console.log("Dialog result: " + result);
+                if (result == true) {
+                    this.advertServ.deleteChat(tappedActivechatItem.chatid);
+                    
+                }
+                else {
+                    this.deleteChat = false;
+                }
+            });
+            
+        
         }
         else{
         const tappedActivechatItem = args.view.bindingContext;
